requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.0
supabase==2.5.0
sentence-transformers==3.0.1
numpy==1.26.4
psycopg2-binary==2.9.9
email-validator==2.1.1
pydantic==2.7.1
python-multipart==0.0.9